# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    Standalone BitTorrent streaming engine for Potok: torrent metadata, direct streaming, HLS, subtitles, thumbnails, and an optional management web UI.
    The plugin/streaming routes are open (players hit them without credentials); the /api/manage/* contour is mounted only with TORRENTGO_ENABLE_WEBUI=true and protected by HTTP Basic auth. Errors are text/plain (via http.Error) unless noted otherwise.
  title: Potok TorrentGo API
  version: "2.0"
paths:
...
  /api/torrents/{hash}/files/{fileIndex}/subtitles/{trackIndex}:
    get:
      description: 'Demuxes an embedded text subtitle track (0-based trackIndex =
        subtitle relIndex from /metadata) and serves it as WebVTT (default), ASS/SSA
        or SRT depending on `format`. With `start` the extraction is windowed around
        the playhead; a window whose pieces are not downloaded yet returns 202 + `Retry-After:
        2` and the client retries as playback nears. Responses carry a content ETag
        — conditional re-requests get 304. Also mounted at the legacy path /stream/{hash}/{fileIndex}/subtitles/{trackIndex}.'
      parameters:
      - description: Infohash (40-char hex)
        in: path
        name: hash
        required: true
        type: string
      - description: File index in the torrent (1-based)
        in: path
        name: fileIndex
        required: true
        type: integer
      - description: Subtitle track index (0-based, per-file subtitle relIndex)
        in: path
        name: trackIndex
        required: true
        type: integer
      - description: 'Output format: webvtt (default) | ass | srt'
        in: query
        name: format
        type: string
      - description: Playhead seconds — extract only the windowed time-slice around
          it
        in: query
        name: start
        type: number
      produces:
      - text/plain
      responses:
        "200":
          description: Subtitle body (text/vtt, text/x-ssa or text/srt)
          schema:
            type: string
        "202":
          description: Window not downloaded yet — retry after Retry-After seconds
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              type: integer
          schema:
            type: string
        "304":
          description: Not modified (matching ETag)
          schema:
            type: string
        "500":
          description: subtitle extraction failed
          schema:
            type: string
      security: []
      summary: Extract an embedded subtitle track
      tags:
      - Media
...
  /stream/{hash}/{fileIndex}/subtitles/{trackIndex}:
    get:
      description: 'Demuxes an embedded text subtitle track (0-based trackIndex =
        subtitle relIndex from /metadata) and serves it as WebVTT (default), ASS/SSA
        or SRT depending on `format`. With `start` the extraction is windowed around
        the playhead; a window whose pieces are not downloaded yet returns 202 + `Retry-After:
        2` and the client retries as playback nears. Responses carry a content ETag
        — conditional re-requests get 304. Also mounted at the legacy path /stream/{hash}/{fileIndex}/subtitles/{trackIndex}.'
      parameters:
      - description: Infohash (40-char hex)
        in: path
        name: hash
        required: true
        type: string
      - description: File index in the torrent (1-based)
        in: path
        name: fileIndex
        required: true
        type: integer
      - description: Subtitle track index (0-based, per-file subtitle relIndex)
        in: path
        name: trackIndex
        required: true
        type: integer
      - description: 'Output format: webvtt (default) | ass | srt'
        in: query
        name: format
        type: string
      - description: Playhead seconds — extract only the windowed time-slice around
          it
        in: query
        name: start
        type: number
      produces:
      - text/plain
      responses:
        "200":
          description: Subtitle body (text/vtt, text/x-ssa or text/srt)
          schema:
            type: string
        "202":
          description: Window not downloaded yet — retry after Retry-After seconds
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              type: integer
          schema:
            type: string
        "304":
          description: Not modified (matching ETag)
          schema:
            type: string
        "500":
          description: subtitle extraction failed
          schema:
            type: string
      security: []
      summary: Extract an embedded subtitle track
      tags:
      - Media
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"